+++ app/classes/shared/file_helpers.py
@@ -27,10 +27,15 @@ class FileHelpers:
                 FileHelpers.del_dirs(sub)
             else:
                 # Delete file if it is a file:
-                sub.unlink()
-
-        # This removes the top-level folder:
-        path.rmdir()
+                try:
+                    sub.unlink()
+                except:
+                    logger.error(f"Unable to delete file {sub}")
+        try:
+            # This removes the top-level folder:
+            path.rmdir()
+        except:
+            logger.error("Unable to remove top level")
         return True
 
     @staticmethod