+++ app/classes/web/panel_handler.py
@@ -453,8 +453,8 @@ class PanelHandler(BaseHandler):
                 for server in un_used_servers[:]:
                     if flag == 0:
                         server["stats"][
-                            "downloading"
-                        ] = self.controller.servers.get_download_status(
+                            "importing"
+                        ] = self.controller.servers.get_import_status(
                             str(server["stats"]["server_id"]["server_id"])
                         )
                         server["stats"]["crashed"] = self.controller.servers.is_crashed(
@@ -572,11 +572,11 @@ class PanelHandler(BaseHandler):
                     "started": "False",
                 }
             if not self.failed_server:
-                page_data["downloading"] = self.controller.servers.get_download_status(
+                page_data["importing"] = self.controller.servers.get_import_status(
                     server_id
                 )
             else:
-                page_data["downloading"] = False
+                page_data["importing"] = False
             page_data["server_id"] = server_id
             try:
                 page_data["waiting_start"] = self.controller.servers.get_waiting_start(
@@ -866,6 +866,18 @@ class PanelHandler(BaseHandler):
                         page_data["users"] = self.controller.users.get_all_users()
                         page_data["roles"] = self.controller.roles.get_all_roles()
                         page_data["auth-servers"][user.user_id] = super_auth_servers
+                        page_data["managed_users"] = []
+            else:
+                page_data["managed_users"] = self.controller.users.get_managed_users(
+                    exec_user["user_id"]
+                )
+                page_data["assigned_roles"] = []
+                for item in page_data["roles"]:
+                    page_data["assigned_roles"].append(item.role_id)
+
+                page_data["managed_roles"] = self.controller.users.get_managed_roles(
+                    exec_user["user_id"]
+                )
 
             template = "panel/panel_config.html"
 
@@ -891,7 +903,7 @@ class PanelHandler(BaseHandler):
                 )
                 return
 
-            page_data["roles_all"] = self.controller.roles.get_all_roles()
+            page_data["roles"] = self.controller.roles.get_all_roles()
             page_data["servers"] = []
             page_data["servers_all"] = self.controller.servers.get_all_defined_servers()
             page_data["role-servers"] = []
@@ -910,8 +922,16 @@ class PanelHandler(BaseHandler):
             )
             if superuser:
                 page_data["super-disabled"] = ""
+                page_data["users"] = self.controller.users.get_all_users()
             else:
                 page_data["super-disabled"] = "disabled"
+
+            page_data["exec_user"] = exec_user["user_id"]
+
+            page_data["manager"] = {
+                "user_id": -100,
+                "username": "None",
+            }
             for file in sorted(
                 os.listdir(os.path.join(self.helper.root_dir, "app", "translations"))
             ):
@@ -1080,9 +1100,21 @@ class PanelHandler(BaseHandler):
             page_data["user"] = self.controller.users.get_user_by_id(user_id)
             page_data["servers"] = set()
             page_data["role-servers"] = page_role_servers
-            page_data["roles_all"] = self.controller.roles.get_all_roles()
+            page_data["roles"] = self.controller.roles.get_all_roles()
+            page_data["exec_user"] = exec_user["user_id"]
             page_data["servers_all"] = self.controller.servers.get_all_defined_servers()
             page_data["superuser"] = superuser
+            if page_data["user"]["manager"] is not None:
+                page_data["manager"] = self.controller.users.get_user_by_id(
+                    page_data["user"]["manager"]
+                )
+            else:
+                page_data["manager"] = {
+                    "user_id": -100,
+                    "username": "None",
+                }
+            if exec_user["superuser"]:
+                page_data["users"] = self.controller.users.get_all_users()
             page_data[
                 "permissions_all"
             ] = self.controller.crafty_perms.list_defined_crafty_permissions()
@@ -1121,6 +1153,17 @@ class PanelHandler(BaseHandler):
                         "/panel/error?error=Unauthorized access: not a user editor"
                     )
                     return
+            if (
+                (
+                    self.controller.users.get_user_by_id(user_id)["manager"]
+                    != exec_user["user_id"]
+                )
+                and not exec_user["superuser"]
+                and str(exec_user["user_id"]) != str(user_id)
+            ):
+                self.redirect(
+                    "/panel/error?error=Unauthorized access: you cannot edit this user"
+                )
 
                 page_data["servers"] = []
                 page_data["role-servers"] = []
@@ -1218,6 +1261,11 @@ class PanelHandler(BaseHandler):
                 defined_servers = self.controller.servers.get_authorized_servers(
                     exec_user["user_id"]
                 )
+
+            page_data["role_manager"] = {
+                "user_id": -100,
+                "username": "None",
+            }
             page_servers = []
             for server in defined_servers:
                 if server not in page_servers:
@@ -1235,6 +1283,7 @@ class PanelHandler(BaseHandler):
             user_roles = self.get_user_roles()
             page_data["new_role"] = False
             role_id = self.get_argument("id", None)
+            role = self.controller.roles.get_role(role_id)
             page_data["role"] = self.controller.roles.get_role_with_servers(role_id)
             if exec_user["superuser"]:
                 defined_servers = self.controller.servers.list_defined_servers()
@@ -1258,7 +1307,21 @@ class PanelHandler(BaseHandler):
             page_data["user-roles"] = user_roles
             page_data["users"] = self.controller.users.get_all_users()
 
-            if EnumPermissionsCrafty.ROLES_CONFIG not in exec_user_crafty_permissions:
+            if page_data["role"]["manager"] is not None:
+                page_data["role_manager"] = self.controller.users.get_user_by_id(
+                    page_data["role"]["manager"]
+                )
+            else:
+                page_data["role_manager"] = {
+                    "user_id": -100,
+                    "username": "None",
+                }
+
+            if (
+                EnumPermissionsCrafty.ROLES_CONFIG not in exec_user_crafty_permissions
+                or exec_user["user_id"] != role["manager"]
+                and not exec_user["superuser"]
+            ):
                 self.redirect(
                     "/panel/error?error=Unauthorized access: not a role editor"
                 )
@@ -1272,8 +1335,15 @@ class PanelHandler(BaseHandler):
         elif page == "remove_role":
             role_id = bleach.clean(self.get_argument("id", None))
 
-            if not superuser:
-                self.redirect("/panel/error?error=Unauthorized access: not superuser")
+            if (
+                not superuser
+                and self.controller.roles.get_role(role_id)["manager"]
+                != exec_user["user_id"]
+            ):
+                self.redirect(
+                    "/panel/error?error=Unauthorized access: not superuser not"
+                    " role manager"
+                )
                 return
             if role_id is None:
                 self.redirect("/panel/error?error=Invalid Role ID")
@@ -1424,6 +1494,7 @@ class PanelHandler(BaseHandler):
                     return
             server_name = self.get_argument("server_name", None)
             server_obj = self.controller.servers.get_server_obj(server_id)
+            shutdown_timeout = self.get_argument("shutdown_timeout", 60)
             if superuser:
                 server_path = self.get_argument("server_path", None)
                 if Helpers.is_os_windows():
@@ -1504,6 +1575,7 @@ class PanelHandler(BaseHandler):
                 )
 
             server_obj.server_name = server_name
+            server_obj.shutdown_timeout = shutdown_timeout
             if superuser:
                 if Helpers.validate_traversal(
                     self.helper.get_servers_root_dir(), server_path
@@ -1936,6 +2008,7 @@ class PanelHandler(BaseHandler):
                     "system user is not editable"
                 )
             user_id = bleach.clean(self.get_argument("id", None))
+            user = self.controller.users.get_user_by_id(user_id)
             username = bleach.clean(self.get_argument("username", None).lower())
             if (
                 username != self.controller.users.get_user_by_id(user_id)["username"]
@@ -1968,7 +2041,19 @@ class PanelHandler(BaseHandler):
             else:
                 superuser = 0
 
-            if not exec_user["superuser"]:
+            if exec_user["superuser"]:
+                manager = self.get_argument("manager")
+                if manager == "":
+                    manager = None
+                else:
+                    manager = int(manager)
+            else:
+                manager = user["manager"]
+
+            if (
+                not exec_user["superuser"]
+                and int(exec_user["user_id"]) != user["manager"]
+            ):
                 if username is None or username == "":
                     self.redirect("/panel/error?error=Invalid username")
                     return
@@ -2019,6 +2104,7 @@ class PanelHandler(BaseHandler):
 
                 user_data = {
                     "username": username,
+                    "manager": manager,
                     "password": password0,
                     "email": email,
                     "enabled": enabled,
@@ -2035,13 +2121,13 @@ class PanelHandler(BaseHandler):
                     user_id, user_data=user_data, user_crafty_data=user_crafty_data
                 )
 
-            self.controller.management.add_to_audit_log(
-                exec_user["user_id"],
-                f"Edited user {username} (UID:{user_id}) with roles {roles} "
-                f"and permissions {permissions_mask}",
-                server_id=0,
-                source_ip=self.get_remote_ip(),
-            )
+                self.controller.management.add_to_audit_log(
+                    exec_user["user_id"],
+                    f"Edited user {username} (UID:{user_id}) with roles {roles} "
+                    f"and permissions {permissions_mask}",
+                    server_id=0,
+                    source_ip=self.get_remote_ip(),
+                )
             self.redirect("/panel/panel_config")
 
         elif page == "edit_user_apikeys":
@@ -2164,6 +2250,15 @@ class PanelHandler(BaseHandler):
             if username is None or username == "":
                 self.redirect("/panel/error?error=Invalid username")
                 return
+
+            if exec_user["superuser"]:
+                manager = self.get_argument("manager")
+                if manager == "":
+                    manager = None
+                else:
+                    manager = int(manager)
+            else:
+                manager = int(exec_user["user_id"])
             # does this user id exist?
             if self.controller.users.get_id_by_name(username) is not None:
                 self.redirect("/panel/error?error=User exists")
@@ -2178,6 +2273,7 @@ class PanelHandler(BaseHandler):
 
             user_id = self.controller.users.add_user(
                 username,
+                manager=manager,
                 password=password0,
                 email=email,
                 enabled=enabled,
@@ -2204,14 +2300,19 @@ class PanelHandler(BaseHandler):
                 server_id=0,
                 source_ip=self.get_remote_ip(),
             )
-            self.controller.crafty_perms.add_user_creation(exec_user["user_id"])
             self.redirect("/panel/panel_config")
 
         elif page == "edit_role":
             role_id = bleach.clean(self.get_argument("id", None))
             role_name = bleach.clean(self.get_argument("role_name", None))
 
-            if EnumPermissionsCrafty.ROLES_CONFIG not in exec_user_crafty_permissions:
+            role = self.controller.roles.get_role(role_id)
+
+            if (
+                EnumPermissionsCrafty.ROLES_CONFIG not in exec_user_crafty_permissions
+                and exec_user["user_id"] != role["manager"]
+                and not exec_user["superuser"]
+            ):
                 self.redirect(
                     "/panel/error?error=Unauthorized access: not a role editor"
                 )
@@ -2227,9 +2328,18 @@ class PanelHandler(BaseHandler):
                 self.redirect("/panel/error?error=Invalid Role ID")
                 return
 
+            if exec_user["superuser"]:
+                manager = self.get_argument("manager", None)
+                if manager == "":
+                    manager = None
+            else:
+                manager = role["manager"]
+
             servers = self.get_role_servers()
 
-            self.controller.roles.update_role_advanced(role_id, role_name, servers)
+            self.controller.roles.update_role_advanced(
+                role_id, role_name, servers, manager
+            )
 
             self.controller.management.add_to_audit_log(
                 exec_user["user_id"],
@@ -2241,6 +2351,12 @@ class PanelHandler(BaseHandler):
 
         elif page == "add_role":
             role_name = bleach.clean(self.get_argument("role_name", None))
+            if exec_user["superuser"]:
+                manager = self.get_argument("manager", None)
+                if manager == "":
+                    manager = None
+            else:
+                manager = exec_user["user_id"]
 
             if EnumPermissionsCrafty.ROLES_CONFIG not in exec_user_crafty_permissions:
                 self.redirect(
@@ -2265,7 +2381,9 @@ class PanelHandler(BaseHandler):
 
             servers = self.get_role_servers()
 
-            role_id = self.controller.roles.add_role_advanced(role_name, servers)
+            role_id = self.controller.roles.add_role_advanced(
+                role_name, servers, manager
+            )
 
             self.controller.management.add_to_audit_log(
                 exec_user["user_id"],
@@ -2273,7 +2391,6 @@ class PanelHandler(BaseHandler):
                 server_id=0,
                 source_ip=self.get_remote_ip(),
             )
-            self.controller.crafty_perms.add_role_creation(exec_user["user_id"])
             self.redirect("/panel/panel_config")
 
         else: